+++ packages/google_maps_flutter/google_maps_flutter_ios_sdk9/CHANGELOG.md
@@ -1,6 +1,7 @@
 ## 2.19.0
 
-* Adds an iOS map-level option to disable implicit marker position and rotation update animations.
+* Adds an iOS-specific map configuration option to disable implicit marker
+  position and rotation update animations.
 
 ## 2.18.8